+++ schedule/schedule-appengine/src/main/webapp/js/controllers.js
@@ -1,0 +1,25 @@
+/**
+ * Created by krico on 02/11/14.
+ */
+var jasifyScheduleControllers = angular.module('jasifyScheduleControllers', ['ui.bootstrap']);
+
+jasifyScheduleControllers.controller('HomeCtrl', ['$scope',
+    function ($scope) {
+    }]);
+
+jasifyScheduleControllers.controller('SignUpCtrl', ['$scope',
+    function ($scope) {
+    }]);
+
+jasifyScheduleControllers.controller('LoginCtrl', ['$scope',
+    function ($scope) {
+    }]);
+
+jasifyScheduleControllers.controller('HelpCtrl', ['$scope',
+    function ($scope) {
+    }]);
+
+jasifyScheduleControllers.controller('ContactUsCtrl', ['$scope',
+    function ($scope) {
+    }]);
+