+++ schedule/schedule-appengine/CHANGELOG.md
@@ -6,6 +6,7 @@ Features:
 
 Changes:
 
+  - installed angular-bootstrap: `bower install angular-bootstrap -S`
   - installed angular-route: `bower install angular-route -S`
   - created login.html
   - changed appengine:devserver to listen on 0.0.0.0 (pom.xml) so I can test templates from the phone.  To get the same